codex refactoring. Reivewing and getting fixed.
