+++ test/AttributedDI.SourceGenerator.UnitTests/Util/AssemblyEmitter.cs
@@ -1,0 +1,28 @@
+using Microsoft.CodeAnalysis;
+using System;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+
+namespace AttributedDI.SourceGenerator.UnitTests.Util;
+
+internal static class AssemblyEmitter
+{
+    public static PortableExecutableReference EmitReference(Compilation compilation, string? assemblyName)
+    {
+        using var stream = new MemoryStream();
+        var emitResult = compilation.Emit(stream);
+        if (!emitResult.Success)
+        {
+            string errorMessages = string.Join(
+                Environment.NewLine,
+                emitResult.Diagnostics
+                    .Where(d => d.Severity == DiagnosticSeverity.Error)
+                    .Select(d => $"  {d.GetMessage(CultureInfo.InvariantCulture)}"));
+            Assert.Fail($"Reference assembly ({assemblyName ?? "Unnamed"}) emit failed:{Environment.NewLine}{errorMessages}");
+        }
+
+        stream.Position = 0;
+        return MetadataReference.CreateFromStream(stream);
+    }
+}
